requirements.txt
Django==5.2.5

djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0

django-cors-headers==4.6.0

psycopg2-binary==2.9.10
dj-database-url==2.1.0

python-decouple==3.8
django-ratelimit==4.1.0

Pillow==11.0.0

gunicorn==21.2.0
whitenoise==6.6.0

django-filter==24.3
